Revert "support "Request.Cancel" for >=go1.7"
