+++ limit/interval_test.go
@@ -24,57 +24,34 @@ func (it *intervalTest) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 }
 
 func TestInterval(t *testing.T) {
-	extInterval := 50 * time.Millisecond
-	interval := NewIntervalTransport(extInterval)
-	interval.GroupKeyFunc = ConstantGroupKeyFunc
-	interval.Transport = new(nilTransport)
-	start := time.Now()
+	it := &intervalTest{}
+	s := httptest.NewServer(it)
+	defer s.Close()
 
-	req, reqErr := http.NewRequest("", "", nil)
-	if reqErr != nil {
-		t.Fatalf("Received unexpected error:\n%v", reqErr)
+	exInterval := 100 * time.Millisecond
+
+	transport := NewIntervalTransport(exInterval)
+	transport.GroupKeyFunc = GroupKeyByHost
+	defer transport.Close()
+
+	testClient := &http.Client{
+		Transport: transport,
 	}
 
-	w := sync.WaitGroup{}
-	w.Add(2)
-
-	{
-		exp := start.Add(extInterval)
+	wg := sync.WaitGroup{}
+	numReq := 10
+	wg.Add(numReq)
+	for i := 0; i < numReq; i++ {
 		go func() {
-			defer w.Done()
-			_, err := interval.RoundTrip(req)
-			if err != nil {
-				t.Fatalf("Received unexpected error:\n%v", err)
-			}
-			act := time.Now()
-
-			dt := exp.Sub(act)
-			if dt < -extInterval || dt > extInterval {
-				t.Fatalf("Max difference between %v and %v allowed is %v, but difference was %v in first response", exp, act, extInterval, dt)
-			}
+			testClient.Get(s.URL)
+			wg.Done()
 		}()
 	}
+	wg.Wait()
 
-	time.Sleep(100 * time.Microsecond) // to ensure request order
-
-	{
-		exp := start.Add(extInterval).Add(extInterval)
-		go func() {
-			defer w.Done()
-			_, err := interval.RoundTrip(req)
-			if err != nil {
-				t.Fatalf("Received unexpected error:\n%v", err)
-			}
-			act := time.Now()
-
-			dt := exp.Sub(act)
-			if dt < -extInterval || dt > extInterval {
-				t.Fatalf("Max difference between %v and %v allowed is %v, but difference was %v in second response", exp, act, extInterval, dt)
-			}
-		}()
+	if it.minInterval < exInterval-(10*time.Millisecond) { // handler may delay
+		t.Errorf("min request interval to server must grater than %s actual %s", exInterval.String(), it.minInterval.String())
 	}
-
-	w.Wait()
 }
 
 func TestIntervalFactory(t *testing.T) {
