+++ limit/rate_limit_base.go
@@ -1,6 +1,7 @@
 package limit
 
 import (
+	"errors"
 	"net/http"
 	"strings"
 	"sync"
@@ -75,6 +76,36 @@ func (t *RateLimit) distKey(req *http.Request) string {
 		return t.GroupKeyFunc(req)
 	}
 	return GroupKeyByHost(req)
+}
+
+// RoundTrip implements the RoundTripper interface.
+func (t *RateLimit) RoundTrip(req *http.Request) (*http.Response, error) {
+	t.initOnce.Do(t.init)
+
+	key := t.distKey(req)
+	if key == "" {
+		return t.transport().RoundTrip(req)
+	}
+
+	reqCh := t.waitCh(key)
+	resCh := make(chan *httpResponseResult)
+	mreq := requestPayload{
+		responder: func() (*http.Response, error) {
+			return t.transport().RoundTrip(req)
+		},
+		resCh: resCh,
+	}
+	select {
+	case t.channelForRequest(reqCh, req) <- mreq:
+		select {
+		case mres := <-resCh:
+			return mres.res, mres.err
+		case <-t.closeCh:
+			return nil, errors.New("request canceled")
+		}
+	case <-t.closeCh:
+		return nil, errors.New("request canceled")
+	}
 }
 
 // CancelRequest cancels an in-flight request by closing its connection.
